+++ engine-dmn/engine/src/test/java/org/operaton/bpm/dmn/engine/impl/hitpolicy/DefaultHitPolicyHandlerRegistryTest.java
@@ -1,0 +1,8 @@
+package org.operaton.bpm.dmn.engine.impl.hitpolicy;
+
+// ...existing imports...
+
+class DefaultHitPolicyHandlerRegistryTest {
+  // TODO: implement tests
+}
+
