--- engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/hitpolicy/HitPolicyTest.priority.compound.dmn
@@ -1,130 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<ns0:definitions xmlns:ns0="https://www.omg.org/spec/DMN/20191111/MODEL/" xmlns:ns1="http://camunda.org/schema/modeler/1.0" xmlns:ns2="http://www.trisotech.com/2015/triso/modeling" xmlns:ns3="https://www.omg.org/spec/DMN/20191111/DMNDI/" xmlns:ns4="http://www.omg.org/spec/DMN/20180521/DC/" xmlns:ns5="http://www.omg.org/spec/DMN/20180521/DI/" id="Definitions_0q20tej" name="DRD" namespace="http://camunda.org/schema/1.0/dmn" exporter="Camunda Modeler" exporterVersion="5.34.0" ns1:executionPlatform="Camunda Platform" ns1:executionPlatformVersion="7.23.0">
-  <ns0:decision id="_3b2953a3-745f-4d2e-b55d-75c8c5ae653c" name="Approval Status">
-    <ns0:variable id="_3cf2bd40-2c83-4e6f-8039-81b391eebf2d" name="Approval Status" typeRef="tApprovalStatus6" />
-    <ns0:informationRequirement id="a4969753-18cf-42d7-ac6d-3821564d2620">
-      <ns0:requiredInput href="#_5a4bdb64-f0ef-4978-9e03-6f1ae64a1f17" />
-    </ns0:informationRequirement>
-    <ns0:informationRequirement id="_57960db1-d798-4bad-8c64-f4fc7e82dbf7">
-      <ns0:requiredInput href="#_41effb45-b3c4-46ac-b1da-122b3e428a98" />
-    </ns0:informationRequirement>
-    <ns0:informationRequirement id="fb1a3d38-22bd-4521-a253-a1c7570b8fa7">
-      <ns0:requiredInput href="#_8ff18665-84e9-49f2-a8df-8981b1844549" />
-    </ns0:informationRequirement>
-    <ns0:decisionTable id="_e7995f4c-0575-4579-8119-aa4f5cb262f6" typeRef="tApprovalStatus6" hitPolicy="PRIORITY" outputLabel="Approval Status">
-      <ns0:input id="_f64e79ef-12df-471a-8830-18dd1060e8f0">
-        <ns0:inputExpression typeRef="long">
-          <ns0:text>Age</ns0:text>
-        </ns0:inputExpression>
-      </ns0:input>
-      <ns0:input id="_066896e9-14c6-4684-a6a7-1036057c69ae">
-        <ns0:inputExpression typeRef="string">
-          <ns0:text>RiskCategory</ns0:text>
-        </ns0:inputExpression>
-        <ns0:inputValues ns2:constraintsType="enumeration">
-          <ns0:text>"High", "Low", "Medium"</ns0:text>
-        </ns0:inputValues>
-      </ns0:input>
-      <ns0:input id="_11764a00-cb22-4e5b-8e52-cbcd2a4d3971">
-        <ns0:inputExpression typeRef="boolean">
-          <ns0:text>isAffordable</ns0:text>
-        </ns0:inputExpression>
-      </ns0:input>
-      <ns0:output id="_c8adc49f-998f-48b8-bd79-746d9c3d25b3" name="Approved/Declined" typeRef="string">
-        <ns0:outputValues ns2:constraintsType="enumeration">
-          <ns0:text>"Approved", "Declined"</ns0:text>
-        </ns0:outputValues>
-      </ns0:output>
-      <ns0:output id="_ee202a75-fdc2-43e0-a9aa-c1d3577f8156" name="Rate" typeRef="string" />
-      <ns0:rule id="_d720490a-c6c7-4e48-b003-44d99720c632">
-        <ns0:inputEntry id="_d720490a-c6c7-4e48-b003-44d99720c632-0">
-          <ns0:text>&gt;=18</ns0:text>
-        </ns0:inputEntry>
-        <ns0:inputEntry id="_d720490a-c6c7-4e48-b003-44d99720c632-1">
-          <ns0:text>"Medium","Low"</ns0:text>
-        </ns0:inputEntry>
-        <ns0:inputEntry id="_d720490a-c6c7-4e48-b003-44d99720c632-2">
-          <ns0:text>true</ns0:text>
-        </ns0:inputEntry>
-        <ns0:outputEntry id="_d720490a-c6c7-4e48-b003-44d99720c632-3">
-          <ns0:text>"Approved"</ns0:text>
-        </ns0:outputEntry>
-        <ns0:outputEntry id="_f269bc24-76ac-4ddb-9b1e-3ab918d32b2f">
-          <ns0:text>"Basic"</ns0:text>
-        </ns0:outputEntry>
-      </ns0:rule>
-      <ns0:rule id="_8e20e9ca-e276-4c26-b09a-52fe55f7b18a">
-        <ns0:inputEntry id="_8e20e9ca-e276-4c26-b09a-52fe55f7b18a-0">
-          <ns0:text>&lt;18</ns0:text>
-        </ns0:inputEntry>
-        <ns0:inputEntry id="_8e20e9ca-e276-4c26-b09a-52fe55f7b18a-1">
-          <ns0:text>-</ns0:text>
-        </ns0:inputEntry>
-        <ns0:inputEntry id="_8e20e9ca-e276-4c26-b09a-52fe55f7b18a-2">
-          <ns0:text>-</ns0:text>
-        </ns0:inputEntry>
-        <ns0:outputEntry id="_8e20e9ca-e276-4c26-b09a-52fe55f7b18a-3">
-          <ns0:text>"Declined"</ns0:text>
-        </ns0:outputEntry>
-        <ns0:outputEntry id="_192d4141-059c-464e-8d84-6f08baded65e">
-          <ns0:text>"Standard"</ns0:text>
-        </ns0:outputEntry>
-      </ns0:rule>
-      <ns0:rule id="_784c2e00-6d8d-4739-8aff-4bd810e4f1d0">
-        <ns0:inputEntry id="_784c2e00-6d8d-4739-8aff-4bd810e4f1d0-0">
-          <ns0:text>-</ns0:text>
-        </ns0:inputEntry>
-        <ns0:inputEntry id="_784c2e00-6d8d-4739-8aff-4bd810e4f1d0-1">
-          <ns0:text>"High"</ns0:text>
-        </ns0:inputEntry>
-        <ns0:inputEntry id="_784c2e00-6d8d-4739-8aff-4bd810e4f1d0-2">
-          <ns0:text>-</ns0:text>
-        </ns0:inputEntry>
-        <ns0:outputEntry id="_784c2e00-6d8d-4739-8aff-4bd810e4f1d0-3">
-          <ns0:text>"Approved"</ns0:text>
-        </ns0:outputEntry>
-        <ns0:outputEntry id="_00de6e01-f0d6-4f82-966c-e09f589b9586">
-          <ns0:text>"Standard"</ns0:text>
-        </ns0:outputEntry>
-      </ns0:rule>
-    </ns0:decisionTable>
-  </ns0:decision>
-  <ns0:inputData id="_41effb45-b3c4-46ac-b1da-122b3e428a98" name="Age">
-    <ns0:variable id="_78044e5c-314d-4029-8a54-573107f67895" name="Age" typeRef="number" />
-  </ns0:inputData>
-  <ns0:inputData id="_5a4bdb64-f0ef-4978-9e03-6f1ae64a1f17" name="RiskCategory">
-    <ns0:variable id="_01068460-8306-4103-b114-be7c0fc45d42" name="RiskCategory" typeRef="string" />
-  </ns0:inputData>
-  <ns0:inputData id="_8ff18665-84e9-49f2-a8df-8981b1844549" name="isAffordable">
-    <ns0:variable id="_57bada07-ae1f-4c31-bb2e-6b536918af85" name="isAffordable" typeRef="boolean" />
-  </ns0:inputData>
-  <ns3:DMNDI>
-    <ns3:DMNDiagram id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_D1">
-      <ns3:Size width="879" height="650" />
-      <ns3:DMNShape id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_s1" dmnElementRef="_3b2953a3-745f-4d2e-b55d-75c8c5ae653c">
-        <ns4:Bounds height="61" width="154" x="362" y="150" />
-      </ns3:DMNShape>
-      <ns3:DMNShape id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_s2" dmnElementRef="_41effb45-b3c4-46ac-b1da-122b3e428a98">
-        <ns4:Bounds height="60" width="153" x="150" y="331" />
-      </ns3:DMNShape>
-      <ns3:DMNShape id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_s3" dmnElementRef="_5a4bdb64-f0ef-4978-9e03-6f1ae64a1f17">
-        <ns4:Bounds height="60" width="153" x="363" y="331" />
-      </ns3:DMNShape>
-      <ns3:DMNShape id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_s4" dmnElementRef="_8ff18665-84e9-49f2-a8df-8981b1844549">
-        <ns4:Bounds height="60" width="153" x="576" y="331" />
-      </ns3:DMNShape>
-      <ns3:DMNEdge id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_e1" dmnElementRef="a4969753-18cf-42d7-ac6d-3821564d2620">
-        <ns5:waypoint x="439" y="331" />
-        <ns5:waypoint x="439" y="211" />
-      </ns3:DMNEdge>
-      <ns3:DMNEdge id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_e2" dmnElementRef="_57960db1-d798-4bad-8c64-f4fc7e82dbf7">
-        <ns5:waypoint x="226" y="331" />
-        <ns5:waypoint x="439" y="211" />
-      </ns3:DMNEdge>
-      <ns3:DMNEdge id="_791b8e95-b7a7-40e7-9dd1-5ff12364f340_e3" dmnElementRef="fb1a3d38-22bd-4521-a253-a1c7570b8fa7">
-        <ns5:waypoint x="652" y="331" />
-        <ns5:waypoint x="439" y="211" />
-      </ns3:DMNEdge>
-    </ns3:DMNDiagram>
-  </ns3:DMNDI>
-</ns0:definitions>
