+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/CollectCountHitPolicyHandlerTest.collect.count.compound.dmn
