+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/CollectHitPolicyHandlerTest.collect.compound.dmn
