+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/CollectMaxHitPolicyHandlerTest.collect.max.compound.dmn
