+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/CollectMinHitPolicyHandlerTest.collect.min.compound.dmn
