+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/CollectSumHitPolicyHandlerTest.collect.sum.single.dmn
