+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/FirstHitPolicyHandlerTest.first.compound.dmn
