+++ engine-dmn/engine/src/test/resources/org/operaton/bpm/dmn/engine/impl/hitpolicy/RuleOrderHitPolicyHandlerTest.ruleOrder.single.dmn
